+++ JudoPayDotNet/Models/CardPaymentModel.cs
@@ -1,3 +1,4 @@
+using System;
 using System.Runtime.Serialization;
 
 namespace JudoPayDotNet.Models
@@ -44,6 +45,12 @@ namespace JudoPayDotNet.Models
         /// </value>
         [DataMember(IsRequired = false)]
         public string IssueNumber { get; set; }
+
+        /// <summary>
+        /// Set true for consumer initiated preAuths that can be incremented before they are fully collected
+        /// </summary>
+        [DataMember(IsRequired = false, EmitDefaultValue = false)]
+        public bool? AllowIncrement { get; set; }
     }
     // ReSharper restore UnusedAutoPropertyAccessor.Global
 }