+++ CHANGELOG.md
@@ -2,7 +2,13 @@
 All notable changes to this project will be documented in this file.
 
 ## 4.2 Changes on 2024-TBD
-- TBD
+- Update Api-Version to 6.21
+- Add optional AllowIncrement boolean flag to CardPaymentModel for CIT preAuths that can be incremented before they are fully collected
+- Add optional AllowIncrement boolean flag to WebPaymentRequestModel for payment sessions created through JudoPayDotNet.Clients.WebPayments.IPreAuths (should not be specified on IPayments or ICheckCards)
+- Add AllowIncrement to GetWebPaymentResponseModel
+- Add JudoPayDotNet.Clients.IPreAuths.IncrementAuth method to increase the authorised amount for an existing CIT preAuth that was set with AllowIncrement=true
+- Add AllowIncrement to PaymentReceiptModel, true on receipts for CIT preAuths that have that flag set to true in the request
+- Add IsIncrementalAuth to PaymentReceiptMode, true on receipts for call to incrementalAuth endpoint
 
 ## 4.1 Changes on 2023-09
 - Update Api-Version to 6.20